fix: 型がfloatのままだったのを修正

+++ Src/Math/Constant.h
@@ -5,9 +5,9 @@ namespace constant {
 #define CONSTANT_G  9.80665
 #define CONSTANT_PI 3.141592653589793 // acos(-1)
 
-    constexpr float fG  = static_cast<float>(CONSTANT_G);
-    constexpr float dG  = static_cast<double>(CONSTANT_G);
-    constexpr float fPI = static_cast<float>(CONSTANT_PI);
-    constexpr float dPI = static_cast<double>(CONSTANT_PI);
+    constexpr float  fG  = static_cast<float>(CONSTANT_G);
+    constexpr double dG  = static_cast<double>(CONSTANT_G);
+    constexpr float  fPI = static_cast<float>(CONSTANT_PI);
+    constexpr double dPI = static_cast<double>(CONSTANT_PI);
 
 }
